EC-513 Separated indev preview into separate file
- added tests for Document & DocumentPreview
