+++ Comments/ClientApp/src/components/App/App.js
@@ -8,8 +8,8 @@ import DocumentViewWithRouter from "../DocumentView/DocumentView";
 import NotFound from "../NotFound/NotFound";
 import ReviewPageWithRouter from "../ReviewPage/ReviewPage";
 import UserProviderWithRouter from "../../context/UserContext";
-import InDevDocumentPreviewWithRouter from "../InDevDocumentPreview/InDevDocumentPreview";
-import InDevDocumentPreviewRedirectWithRouter from "../InDevDocumentPreview/InDevDocumentPreviewRedirect";
+import DocumentPreviewWithRouter from "../DocumentPreview/DocumentPreview";
+import DocumentPreviewRedirectWithRouter from "../DocumentPreview/DocumentPreviewRedirect";
 
 type PropsType = any;
 
@@ -27,25 +27,23 @@ class App extends React.Component<PropsType, StateType> {
 
 				<Switch>
 					{/*Document View*/}
-					<Route path="/:consultationId/:documentId/:chapterSlug">
+					<Route exact path="/:consultationId/:documentId/:chapterSlug">
 						<DocumentViewWithRouter/>
 					</Route>
 
-					<Switch>
-						{/*Document (Preview Layout)*/}
-						<Route path="/preview/:reference/consultation/:consultationId/document/:documentId/chapter/:chapterSlug">
-							<InDevDocumentPreviewWithRouter />
-						</Route>
+					{/*Document (Preview Layout)*/}
+					<Route exact path="/preview/:reference/consultation/:consultationId/document/:documentId/chapter/:chapterSlug">
+						<DocumentPreviewWithRouter />
+					</Route>
 
-						{/*	If we hit this we're coming in *without* a chapter slug, so we need to get the first chapter of the current document and pass its slug into the URL so it matches the route above */}
-						<Route path="/preview/:reference/consultation/:consultationId/document/:documentId">
-							<InDevDocumentPreviewRedirectWithRouter />
-							{/* This component only redirects to the above route */}
-						</Route>
-					</Switch>
+					{/*	If we hit this we're coming in *without* a chapter slug, so we need to get the first chapter of the current document and pass its slug into the URL so it matches the route above */}
+					<Route exact path="/preview/:reference/consultation/:consultationId/document/:documentId">
+						<DocumentPreviewRedirectWithRouter />
+						{/* This component only redirects to the above route */}
+					</Route>
 
 					{/*Review Page*/}
-					<Route path="/:consultationId/review">
+					<Route exact path="/:consultationId/review">
 						<ReviewPageWithRouter/>
 					</Route>
