+++ Comments/ClientApp/src/components/DocumentPreview/__tests__/PreviewChapter.json
@@ -1,0 +1,6 @@
+{
+  "content": "<div class=\"chapter\" title=\"DRAFT_Appendices_A-T_CLEAN.docx\" xmlns=\"http://www.w3.org/1999/xhtml\">\r\n  <h2 class=\"title\">\r\n    <a id=\"draftappendicesa-tcleandocx\" style=\"position:relative\" class=\"annotator-chapter\" data-heading-type=\"chapter\" title=\"Comment on chapter\" xmlns=\"\">DRAFT_Appendices_A-T_CLEAN.docx<span class=\"annotator-adder\" /></a>\r\n    <!--[E] - Error: This document contains neither a recognisable issue date nor a recognisable revision date, so appears to be completely undated-->\r\n    <!--[E] - Error: No recognisable issue date found-->\r\n    <!--[E] - Error: No main title found-->\r\n    <!--[E] - Error: This document contains neither a recognisable issue date nor a recognisable revision date, so appears to be completely undated-->\r\n    <!--[E] - Error: No recognisable issue date found-->\r\n    <!--[E] - Error: No main title found-->\r\n    <!--[E] - Error: This document contains neither a recognisable issue date nor a recognisable revision date, so appears to be completely undated-->\r\n    <!--[E] - Error: No recognisable issue date found-->\r\n    <!--[E] - Error: No main title found-->\r\n    <!--[E] - Error: This document contains neither a recognisable issue date nor a recognisable revision date, so appears to be completely undated-->\r\n    <!--[E] - Error: No recognisable issue date found-->\r\n    <!--[E] - Error: No main title found-->\r\n    <!--[E] - Error: This document contains neither a recognisable issue date nor a recognisable revision date, so appears to be completely undated-->\r\n    <!--[E] - Error: No recognisable issue date found-->\r\n    <!--[E] - Error: No main title found-->\r\n    <!--[E] - Error: This document contains neither a recognisable issue date nor a recognisable revision date, so appears to be completely undated-->\r\n    <!--[E] - Error: No recognisable issue date found-->\r\n    <!--[E] - Error: No main title found-->\r\n    <!--[E] - Error: This document contains neither a recognisable issue date nor a recognisable revision date, so appears to be completely undated-->\r\n    <!--[E] - Error: No recognisable issue date found-->\r\n    <!--[E] - Error: No main title found-->\r\n    <!--[W] - Warning: Text found before first section heading, so software inserted section heading with title 'DRAFT_Appendices_A-T_CLEAN.docx'-->\r\n  </h2>\r\n  <p></p>\r\n  <p></p>\r\n</div>",
+  "sections": [],
+  "slug": "draftappendicesa-tcleandocx",
+  "title": "DRAFT_Appendices_A-T_CLEAN.docx"
+}
