+++ Comments/ClientApp/src/components/ReviewPage/ReviewPage.js
@@ -61,7 +61,7 @@ export class ReviewPage extends Component<PropsType, StateType> {
 			viewSubmittedComments: false,
 			validToSubmit: false,
 			shouldShowCommentsTab: true,
-			shouldShowQuestionsTab: true
+			shouldShowQuestionsTab: true,
 		};
 	}
 
@@ -75,7 +75,7 @@ export class ReviewPage extends Component<PropsType, StateType> {
 			});
 
 		const consultationData = load("consultation", undefined, [], {
-			consultationId, isReview: true
+			consultationId, isReview: true,
 		})
 			.then(response => response.data)
 			.catch(err => {
@@ -84,7 +84,7 @@ export class ReviewPage extends Component<PropsType, StateType> {
 
 		return {
 			documentsList: await documentsData,
-			consultationData: await consultationData
+			consultationData: await consultationData,
 		};
 	};
 
@@ -96,14 +96,14 @@ export class ReviewPage extends Component<PropsType, StateType> {
 						label: consultationDocument.title,
 						url: `${this.props.location.pathname}?sourceURI=${encodeURIComponent(consultationDocument.sourceURI)}`,
 						current: this.getCurrentSourceURI() === consultationDocument.sourceURI,
-						isReactRoute: true
+						isReactRoute: true,
 					};
 				}
 			);
 
 		return {
 			title: "View comments by document",
-			links: documentLinks
+			links: documentLinks,
 		};
 	};
 
@@ -112,10 +112,10 @@ export class ReviewPage extends Component<PropsType, StateType> {
 		this.gatherData()
 			.then(data => {
 
-				const shouldShowCommentsTab = (	data.consultationData.supportsComments || 
+				const shouldShowCommentsTab = (	data.consultationData.supportsComments ||
 												data.consultationData.consultationState.hasUserSuppliedComments);
 
-				const shouldShowQuestionsTab = (	data.consultationData.consultationState.hasQuestions || 
+				const shouldShowQuestionsTab = (	data.consultationData.consultationState.hasQuestions ||
 													data.consultationData.consultationState.hasUserSuppliedAnswers);
 
 				this.setState({
@@ -123,7 +123,7 @@ export class ReviewPage extends Component<PropsType, StateType> {
 					userHasSubmitted: data.consultationData.consultationState.userHasSubmitted,
 					validToSubmit: data.consultationData.consultationState.supportsSubmission,
 					shouldShowCommentsTab,
-					shouldShowQuestionsTab
+					shouldShowQuestionsTab,
 				});
 			})
 			.catch(err => {
@@ -149,23 +149,23 @@ export class ReviewPage extends Component<PropsType, StateType> {
 		let answersToSubmit = [];
 
 		if (typeof(this.questionList) !== "undefined"){
-			const questions = this.questionList.getQuestions();		
+			const questions = this.questionList.getQuestions();
 			questions.forEach(function(question){
 				if (question.answers != null){
 					answersToSubmit = answersToSubmit.concat(question.answers);
-				}			
+				}
 			});
 		}
 		let commentsAndAnswers = {comments: comments, answers: answersToSubmit};
 
 		load("submit", undefined, [], {}, "POST", commentsAndAnswers, true)
-			.then(res => {
+			.then(() => {
 				this.submittedHandler();
 			})
 			.catch(err => {
 				console.log(err);
 				if (err.response) alert(err.response.statusText);
-			});		
+			});
 
 	};
 
@@ -173,9 +173,9 @@ export class ReviewPage extends Component<PropsType, StateType> {
 		this.setState({
 			userHasSubmitted: true,
 			validToSubmit: false,
-			viewSubmittedComments: false
+			viewSubmittedComments: false,
 		});
-	}
+	};
 
 	//this validation handler code is going to have to get a bit more advanced when questions are introduced, as it'll be possible
 	//to answer a question on the review page and the submit button should then enable - if the consultation is open + hasn't already been submitted + all the mandatory questions are answered.
@@ -185,7 +185,7 @@ export class ReviewPage extends Component<PropsType, StateType> {
 		let hasAnswers = false;
 		if (typeof(this.questionList) !== "undefined"){
 			const questions = this.questionList.getQuestions();
-			
+
 			questions.forEach(function(question){
 				if (question.answers !== null && question.answers.length > 0){
 					hasAnswers = true;
@@ -193,19 +193,19 @@ export class ReviewPage extends Component<PropsType, StateType> {
 			});
 		}
 		this.setState({
-			validToSubmit: comments.length > 0 || hasAnswers
+			validToSubmit: comments.length > 0 || hasAnswers,
 		});
-	}
+	};
 
 	viewSubmittedCommentsHandler = () => {
 		this.setState({
-			viewSubmittedComments: true
+			viewSubmittedComments: true,
 		});
-	}
+	};
 
 	render() {
 		if (this.state.documentsList.length === 0) return <h1>Loading...</h1>;
-		const { title, reference, endDate } = this.state.consultationData;
+		const { title, reference } = this.state.consultationData;
 
 		if (!this.state.shouldShowQuestionsTab && !this.state.shouldShowCommentsTab){
 			return null; //shouldn't really be able to get here.
@@ -245,9 +245,9 @@ export class ReviewPage extends Component<PropsType, StateType> {
 										</div>
 										:
 										<StickyContainer className="grid">
-											<div data-g="12 md:9 md:push:3">												
+											<div data-g="12 md:9 md:push:3">
 												{/* <div className="tabs" data-tabs> */}
-													{/* <ul className="tabs__list" role="tablist">
+												{/* <ul className="tabs__list" role="tablist">
 														{this.state.shouldShowQuestionsTab &&
 															<li className={`tabs__tab ${this.props.viewComments ? "" : "tabs__tab--active"}`} role="presentation">
 																<button className="tabs__tab-btn" type="button" role="tab">
@@ -268,10 +268,10 @@ export class ReviewPage extends Component<PropsType, StateType> {
 															</button>
 														</li>
 													</ul> */}
-													{/* <div className="tabs__content"> */}
-														{this.state.shouldShowQuestionsTab &&
+												{/* <div className="tabs__content"> */}
+												{this.state.shouldShowQuestionsTab &&
 															<Fragment>
-															{/* <div className="tabs__pane" role="tabpanel"> */}
+																{/* <div className="tabs__pane" role="tabpanel"> */}
 																<h3 className="mt--0" id="comments-column">Questions</h3>
 																<CommentListWithRouter
 																	isReviewPage={true}
@@ -281,12 +281,12 @@ export class ReviewPage extends Component<PropsType, StateType> {
 																	submittedHandler={this.submittedHandler}
 																	validationHander={this.validationHander}
 																	viewComments={false}/>
-															{/* </div> */}
+																{/* </div> */}
 															</Fragment>
-														}
-														{this.state.shouldShowCommentsTab &&
+												}
+												{this.state.shouldShowCommentsTab &&
 															<Fragment>
-															{/* <div className="tabs__pane" role="tabpanel"> */}
+																{/* <div className="tabs__pane" role="tabpanel"> */}
 																<h3 className="mt--0" id="comments-column">Comments</h3>
 																<CommentListWithRouter
 																	isReviewPage={true}
@@ -296,71 +296,71 @@ export class ReviewPage extends Component<PropsType, StateType> {
 																	submittedHandler={this.submittedHandler}
 																	validationHander={this.validationHander}
 																	viewComments={true}/>
-															{/* </div> */}
+																{/* </div> */}
 															</Fragment>
-														}
-														{/* <div className="tabs__pane" role="tabpanel"> */}
-															{this.state.userHasSubmitted ?
-																<div className="hero">
-																	<div className="hero__container">
-																		<div className="hero__body">
-																			<div className="hero__copy">
-																				<p className="hero__intro" data-qa-sel="submitted-text">Thank you, your comments have been submitted.</p>
-																				<button
-																					className="btn btn--secondary">
+												}
+												{/* <div className="tabs__pane" role="tabpanel"> */}
+												{this.state.userHasSubmitted ?
+													<div className="hero">
+														<div className="hero__container">
+															<div className="hero__body">
+																<div className="hero__copy">
+																	<p className="hero__intro" data-qa-sel="submitted-text">Thank you, your comments have been submitted.</p>
+																	<button
+																		className="btn btn--secondary">
 																					Download all comments
-																				</button>
-																			</div>
-																		</div>
-																	</div>
+																	</button>
 																</div>
-																:
-																<div className="hero">
-																	<div className="hero__container">
-																		<div className="hero__body">
-																			<div className="hero__copy">
-																				{/* <h1 className="hero__title">Hero title</h1> */}
-																				<p className="hero__intro">You are about to submit your final response to NICE</p>
-																				<p>After submission you won't be able to:</p>
-																				<ul>
-																					<li>edit your comments further</li>
-																					<li>add any extra comments.</li>
-																				</ul>
-																				<p>Do you want to continue?</p>
-																				{/* <div className="hero__actions">
+															</div>
+														</div>
+													</div>
+													:
+													<div className="hero">
+														<div className="hero__container">
+															<div className="hero__body">
+																<div className="hero__copy">
+																	{/* <h1 className="hero__title">Hero title</h1> */}
+																	<p className="hero__intro">You are about to submit your final response to NICE</p>
+																	<p>After submission you won't be able to:</p>
+																	<ul>
+																		<li>edit your comments further</li>
+																		<li>add any extra comments.</li>
+																	</ul>
+																	<p>Do you want to continue?</p>
+																	{/* <div className="hero__actions">
 																					<a className="btn btn--cta">Yes, submit my response</a>
 																					<a className="btn" target="_blank" rel="noopener external">No, keep editing</a>
 																				</div> */}
-																				<UserContext.Consumer>
-																					{contextValue => {
-																						if (contextValue.isAuthorised) {
-																							return (
-																								<Fragment>
-																									<h3 className="mt--0">Ready to submit?</h3>
-																									<button
-																										disabled={!this.state.validToSubmit}
-																										className="btn btn--cta"
-																										data-qa-sel="submit-comment-button"
-																										onClick={this.submitConsultation}
-																									>
-																										{this.state.userHasSubmitted ? "Comments submitted": "Submit your comments"}
-																									</button>
-																									<button
-																										className="btn btn--secondary">
+																	<UserContext.Consumer>
+																		{contextValue => {
+																			if (contextValue.isAuthorised) {
+																				return (
+																					<Fragment>
+																						<h3 className="mt--0">Ready to submit?</h3>
+																						<button
+																							disabled={!this.state.validToSubmit}
+																							className="btn btn--cta"
+																							data-qa-sel="submit-comment-button"
+																							onClick={this.submitConsultation}
+																						>
+																							{this.state.userHasSubmitted ? "Comments submitted": "Submit your comments"}
+																						</button>
+																						<button
+																							className="btn btn--secondary">
 																										Download all comments
-																									</button>
-																								</Fragment>
-																							);
-																						}
-																					}}
-																				</UserContext.Consumer>
-																			</div>
-																		</div>
-																	</div>
+																						</button>
+																					</Fragment>
+																				);
+																			}
+																		}}
+																	</UserContext.Consumer>
 																</div>
-															}
-														{/* </div> */}
-													{/* </div> */}
+															</div>
+														</div>
+													</div>
+												}
+												{/* </div> */}
+												{/* </div> */}
 												{/* </div> */}
 											</div>
 											<div data-g="12 md:3 md:pull:9">
@@ -375,9 +375,9 @@ export class ReviewPage extends Component<PropsType, StateType> {
 																			label: title,
 																			url: this.props.location.pathname,
 																			current: this.getCurrentSourceURI() == null,
-																			isReactRoute: true
-																		}
-																	]
+																			isReactRoute: true,
+																		},
+																	],
 																}
 															}/>
 															<StackedNav
