+++ Comments/ClientApp/src/components/ReviewPage/__tests__/ReviewPage.test.js
@@ -96,7 +96,7 @@ describe("[ClientApp] ", () => {
 		});
 
 		// it.only("should hit the submit endpoint successfully", async done => {
-					
+
 		// 	const mock = new MockAdapter(axios);
 
 		// 	const wrapper = mount(
@@ -127,16 +127,16 @@ describe("[ClientApp] ", () => {
 		// 		.onPost("/consultations/api/Submit")
 		// 		.reply(() => {
 		// 			done();
-					
+
 		// 		});
-			
+
 		// 	return Promise.all([
 		// 		documentsPromise,
 		// 		consultationPromise
 		// 	]).then(async () => {
 		// 		await nextTick();
 		// 		wrapper.update();
-				
+
 		// 		//expect(wrapper.find(ReviewPage).instance().state.isSubmitted).toEqual(false);
 
 		// 		wrapper.find(ReviewPage).instance().submitConsultation();
@@ -144,7 +144,7 @@ describe("[ClientApp] ", () => {
 
 		// });
 
-		it.only("should match snapshot with supplied data", () => {
+		it("should match snapshot with supplied data", () => {
 			const mock = new MockAdapter(axios);
 
 			const wrapper = mount(
