+++ Comments/ClientApp/src/main.scss
@@ -23,3 +23,4 @@ $nice-font-base-path: "/consultations/fonts/";
 @import "./components/CommentBox/CommentBox";
 @import "./components/PhaseBanner/PhaseBanner";
 @import "./components/Selection/Selection";
+@import "./document-processing/transforms/comment";
