+++ Comments/ClientApp/src/styles/variables.scss
@@ -5,6 +5,10 @@
   box-shadow: -10px 0 10px rgba(0, 0, 0, 0.1);
 }
 
+%all-drop-shadow {
+  box-shadow: 0 0 rem(10px) rgba(0, 0, 0, 0.1);
+}
+
 %transition-default {
   transition: all 0.3s ease-in-out;
   transition-property: transform;
